requirements.txt
selenium
Pillow
requests
beautifulsoup4